Feat: Refactor exercise pages and add modal alerts
This update improves the structure and functionality of the exercise check-game, refactoring and assignment pages while enhancing the user experience with new modal alerts.

Changes include:
- Refactoring the check-game, refactoring and assignment pages to use the reusable components introduced in the previous commit.
- Adding modal alerts for key events:
  - Exercise completion.
  - Exercise failure.
  - Level up.
  - New badge unlock.
- Integrating the modal alerts into the check-game and refactoring pages to provide real-time feedback to users.

These changes improve code modularity, promote reusability across pages and enhance user engagement through interactive modal feedback.
