+++ Test-Smells-Game/Projects/Frontend/src/app/components/alert/achivement-alert/achievement-alert.component.html
@@ -1,0 +1,8 @@
+<div *ngIf="showModal" class="modal-overlay">
+  <div class="modal-content level-up-modal">
+    <h1 class="title">⭐ {{ title | translate }} ⭐</h1>
+    <p class="message">{{ message | translate }}</p>
+    <button class="close-button" (click)="close()">{{ 'Awesome!' | translate}}</button>
+  </div>
+</div>
+
