+++ Test-Smells-Game/Projects/Frontend/src/app/components/alert/success-alert/success-alert.component.ts
@@ -1,0 +1,19 @@
+import {Component, EventEmitter, Input, Output} from '@angular/core';
+
+@Component({
+  selector: 'app-success-alert',
+  templateUrl: './success-alert.component.html',
+  styleUrls: ['./success-alert.component.css']
+})
+export class SuccessAlertComponent {
+  @Input() message: string = "You have completed successfully the exercise.";
+  showModal: boolean = false;
+
+  show(): void {
+    this.showModal = true;
+  }
+
+  close() {
+    this.showModal = false
+  }
+}
