+++ Test-Smells-Game/Projects/Frontend/src/app/routes/assignments-route/assignments-core/assignments-core-route.component.html
@@ -1,0 +1,114 @@
+<div class="container">
+  <div class="assignment-header">
+    <h1>{{ assignmentName }}</h1>
+    <div *ngIf="assignment && assignment.students && assignment.students.length > 0">
+      <p>
+        <strong style="color: #B22222;">{{ 'End time' | translate }}</strong>: <strong style="color: #B22222;">{{ this.currentStudent.endDate }} {{this.currentStudent.endTime}}</strong>
+      </p>
+    </div>
+
+  </div>
+  <div class="note" style="margin-top: 20px;">
+    <p>{{ 'If you don\'t finish by the end time, the assignment will be failed' | translate }}</p>
+  </div>
+  <hr>
+
+  <div *ngIf="assignment?.gameType === 'refactoring'">
+    <div class="container">
+      <h1>{{exerciseName}}</h1>
+      <div class="row">
+        <div class="col">
+        </div>
+        <div class="row">
+          <div class="col">
+            <app-refactoring-editor-card #code [editorHeader]="exerciseName+'.java'" [injectedCode]="refactoringService.userCode" [editable]="false"
+                                         [progressBarMode]="refactoringService.progressBarMode">
+            </app-refactoring-editor-card>
+          </div>
+
+          <mat-divider></mat-divider>
+
+          <div class="col" style="padding-top: 20px">
+            <app-refactoring-editor-card #testing [editorHeader]="exerciseName+'Test.java'" [injectedCode]="refactoringService.testingCode"
+                                         [progressBarMode]="refactoringService.progressBarMode">
+            </app-refactoring-editor-card>
+          </div>
+
+          <mat-divider></mat-divider>
+
+          <div style="margin-top: 20px">
+            <button type="button" class="btn btn-primary" (click)="compile()" [disabled]="refactoringService.progressBarMode == 'query'">{{ 'Compile' | translate }}</button>
+            <button style="margin-left: 20px;" type="button" class="btn btn-primary" [disabled]="submitIsDisabled()" (click)="submitRefactoringAssignment()">{{ 'Finish' | translate }}</button>
+          </div>
+
+          <mat-divider></mat-divider>
+
+          <div class="col" style="margin-top: 20px">
+            <app-refactoring-editor-card #output [editorHeader]="'Shell'" [injectedCode]="refactoringService.shellCode" [editable]="false"
+                                         [progressBarMode]="refactoringService.progressBarMode">
+            </app-refactoring-editor-card>
+          </div>
+
+          <app-refactoring-compilation-results></app-refactoring-compilation-results>
+
+          <div *ngIf="refactoringService.exerciseIsCompiledSuccessfully">
+            <app-refactoring-score
+              [score]="Math.abs(refactoringService.smellNumber - refactoringService.exerciseConfiguration.refactoring_game_configuration.smells_allowed)"
+              [smellNumber]="refactoringService.smellNumber"
+              [refactoringResult]="refactoringService.refactoringResult"
+              [originalCoverage]="refactoringService.originalCoverage"
+              [refactoredCoverage]="refactoringService.refactoredCoverage"
+              [smellNumberWarning]="refactoringService.smellNumberWarning"
+              [exerciseConfiguration]="refactoringService.exerciseConfiguration"
+            ></app-refactoring-score>
+
+            <app-refactoring-smell-list
+              [smellList]="refactoringService.smellList"
+              [smellDescriptions]="refactoringService.smellDescriptions"
+              [methodList]="refactoringService.methodList"
+            ></app-refactoring-smell-list>
+          </div>
+
+          <button type="button" class="btn btn-primary" routerLink="/assignment/leaderboard/{{exerciseName}}">{{ 'Solution Repository' | translate }}</button>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div *ngIf="assignment?.gameType === 'check-smell'">
+    <div class="container">
+      <h1>{{exerciseName}}</h1>
+      <div>
+        <div *ngIf="!checkSmellService.exerciseCompleted else showResult" class="card">
+
+          <app-check-smell-question #testing [editorHeader]="exerciseName+'.java'" [injectedCode]="checkSmellService.questions[checkSmellService.actualQuestionNumber].questionCode" [editable]="false">
+          </app-check-smell-question>
+
+          <app-check-smell-answer-list [questionAnswers]="checkSmellService.questions[checkSmellService.actualQuestionNumber].answers">
+          </app-check-smell-answer-list>
+
+          <div>
+            <button class="btn btn-primary" *ngIf="checkSmellService.actualQuestionNumber !== 0" (click)="checkSmellService.goBackward()">{{ 'Back' | translate }}</button>
+            <button class="btn btn-primary" *ngIf="checkSmellService.actualQuestionNumber < checkSmellService.questions.length - 1" (click)="checkSmellService.goForward()">{{ 'Next' | translate }}</button>
+          </div>
+          <button *ngIf="checkSmellService.renderResultsButton()" class="btn btn-primary" (click)="this.submitExercise()">{{ 'Submit' | translate }}</button>
+
+        </div>
+      </div>
+
+      <ng-template #showResult>
+        <h1>{{ 'Score' | translate }}: {{Math.round((checkSmellService.score * 100) / checkSmellService.assignmentScore) + '/100'}}</h1>
+        <div *ngFor="let question of checkSmellService.questions" class="card">
+
+          <app-check-smell-question [injectedCode]="question.questionCode">
+          </app-check-smell-question>
+
+          <app-check-smell-answer-list [questionAnswers]="checkSmellService.questions[checkSmellService.actualQuestionNumber].answers"
+                                       [showBackground]="true">
+          </app-check-smell-answer-list>
+        </div>
+      </ng-template>
+
+    </div>
+  </div>
+</div>
