+++ Test-Smells-Game/Projects/Frontend/src/app/routes/assignments-route/assignments-core/assignments-core-route.component.spec.ts
