+++ Test-Smells-Game/Projects/Frontend/src/app/routes/assignments-route/assignments-list/assignments-list-route.component.html
@@ -7,36 +7,36 @@
   <h2 style="color: red; font-size: 20px; text-align: center;">{{ "Please contact your supervisor" | translate }}</h2>
 </div>
 
-<div *ngIf="checkAssignmentPresence() === 0 && errorMessage === undefined">
+<div *ngIf="!(hasStudentAssignments('refactoring') || hasStudentAssignments('check-smell')) && errorMessage === undefined">
   <br><br><br><br><br><br><br><br>
   <h1 style="font-size: 20px; text-align: center;">
     {{ 'There are currently no assignments you are participating in' | translate }}
   </h1>
 </div>
 
-<h1 style="font-size: 20px; text-align: center;">
-  {{ 'Refactoring type assignments' | translate }}
-</h1>
-<div *ngFor="let assignment of filterAssignmentsByGameType('refactoring')">
-  <div *ngIf="checkAssignmentPresence() > 0" style="border: 1px solid #ccc; border-radius: 8px; padding: 16px; margin-bottom: 16px;">
+<div *ngIf="hasStudentAssignments('refactoring')">
+  <h1 style="font-size: 20px; text-align: center;">
+    {{ 'Refactoring type assignments' | translate }}
+  </h1>
+  <div *ngFor="let assignment of filterAssignmentsByGameType('refactoring')" style="border: 1px solid #ccc; border-radius: 8px; padding: 16px; margin-bottom: 16px;">
     <h2 style="font-size: 24px; margin-bottom: 8px;">{{ assignment.name }}</h2>
     <p style="font-size: 16px; color: #666; margin-bottom: 12px;"> {{ 'Assignment type' | translate }}: {{ assignment.gameType }}</p>
 
     <ul style="list-style-type: none; padding: 0;">
       <ng-container *ngFor="let student of assignment.students">
         <ng-container *ngIf="student.name === currentUser?.userName">
-          <ng-container *ngIf="assignment.type === 'competitivo' && student.submitted === true">
+          <ng-container *ngIf="assignment.type === 'competitivo' && student.submitted">
             <li style="margin-bottom: 8px;">
               <span style="color: red;">{{ 'You have already completed the assignment' | translate }}</span>
             </li>
           </ng-container>
-          <ng-container *ngIf="!(assignment.type === 'competitivo' && student.submitted === true)">
+          <ng-container *ngIf="!(assignment.type === 'competitivo' && student.submitted)">
             <li *ngIf="isAssignmentActive(student.startDate, student.startTime, student.endDate, student.endTime)" style="margin-bottom: 8px;">
               <a [routerLink]="['/assignments', assignment.name, student.exercise]" style="text-decoration: none; color: #007bff;">
                 {{ student.exercise }}
               </a>
               <p style="margin-top: 12px;"> {{ 'Available from' | translate }} {{ student.startDate }} {{ "at" | translate}} {{ student.startTime }} </p>
-              <p> {{ "To be delivered by" | translate }} {{ student.startDate }} {{ "at" | translate}} {{ student.endTime }} </p>
+              <p> {{ "To be delivered by" | translate }} {{ student.endDate }} {{ "at" | translate}} {{ student.endTime }} </p>
             </li>
             <li *ngIf="!isAssignmentActive(student.startDate, student.startTime, student.endDate, student.endTime)" style="margin-bottom: 8px;">
               <span style="color: red;">{{ 'The assignment is unavailable' | translate }}</span>
@@ -48,11 +48,11 @@
   </div>
 </div>
 
-<h1 style="font-size: 20px; text-align: center;">
-  {{ 'Check smell type assignments' | translate }}
-</h1>
-<div *ngFor="let assignment of filterAssignmentsByGameType('check-smell')">
-  <div *ngIf="checkAssignmentPresence() > 0" style="border: 1px solid #ccc; border-radius: 8px; padding: 16px; margin-bottom: 16px;">
+<div *ngIf="hasStudentAssignments('check-smell')">
+  <h1 style="font-size: 20px; text-align: center;">
+    {{ 'Check smell type assignments' | translate }}
+  </h1>
+  <div *ngFor="let assignment of filterAssignmentsByGameType('check-smell')" style="border: 1px solid #ccc; border-radius: 8px; padding: 16px; margin-bottom: 16px;">
     <h2 style="font-size: 24px; margin-bottom: 8px;">{{ assignment.name }}</h2>
     <p style="font-size: 16px; color: #666; margin-bottom: 12px;"> {{ 'Assignment type' | translate }}: {{ assignment.gameType }}</p>
