+++ Test-Smells-Game/Projects/Frontend/src/app/routes/assignments-route/assignments-list/assignments-list-route.spec.ts
