--- Test-Smells-Game/Projects/Frontend/src/app/routes/assignments/assignments-core/assignments-core-route.component.html
@@ -1,42 +1,0 @@
-<div class="container">
-  <div class="assignment-header">
-    <h1>{{ assignmentName }}</h1>
-    <div *ngIf="assignment && assignment.students && assignment.students.length > 0">
-      <p>
-        <strong style="color: #B22222;">{{ 'End time' | translate }}</strong>: <strong style="color: #B22222;">{{ this.currentStudent.endDate }} {{this.currentStudent.endTime}}</strong>
-      </p>
-    </div>
-
-  </div>
-  <div class="note" style="margin-top: 20px;">
-    <p>{{ 'If you don\'t finish by the end time, the assignment will be failed' | translate }}</p>
-  </div>
-  <hr>
-
-  <app-refactoring-core *ngIf="assignment?.gameType === 'refactoring'"
-                        [exerciseName]="exerciseName"
-                        #code [userCode]="userCode"
-                        [progressBarMode]="progressBarMode"
-                        #testing [testingCode]="testingCode"
-                        #output [shellCode]="shellCode"
-                        [exerciseSuccess]="exerciseSuccess"
-                        [refactoringWarning]="refactoringWarning"
-                        [smellNumber]="smellNumber"
-                        [refactoringResult]="refactoringResult"
-                        [originalCoverage]="originalCoverage"
-                        [refactoredCoverage]="refactoredCoverage"
-                        [smellNumberWarning]="smellNumberWarning"
-                        [exerciseConfiguration]="exerciseConfiguration"
-                        [smellList]="smellList"
-                        [methodList]="methodList"
-                        [smellDescriptions]="smellDescriptions"
-                        [solutionRepoRoute]="solutionRepoRoute"
-                        [compileCode]="compile"
-                        [submitExercise]="submitRefactoringAssignment"
-                        [submitIsEnabled]="submitIsEnabled"
-                        [finishButtonText]="'Finish'"
-                        [endButtonText]="'Solution Repository'"
-                        />
-
-  <app-check-core *ngIf="assignment?.gameType === 'check-smell'" [handleSubmit]="submitCheckSmellAssignment"/>
-</div>
