+++ Test-Smells-Game/Projects/Frontend/src/app/routes/check-smell-game/check-game-core/check-game-core-route.component.ts
@@ -1,15 +1,13 @@
-import { Component, HostListener, OnInit } from '@angular/core';
-import { CodeeditorService } from "../../../services/codeeditor/codeeditor.service";
-import { ExerciseService } from "../../../services/exercise/exercise.service";
-import { ActivatedRoute } from "@angular/router";
-import { Question } from "../../../model/question/question.model";
-import { Answer } from "../../../model/question/answer.model";
-import { CheckGameExerciseConfig } from "../../../model/exercise/ExerciseConfiguration.model";
-import { User } from "../../../model/user/user.model";
-import { MatCheckbox } from "@angular/material/checkbox";
-import { UserService } from '../../../services/user/user.service';
-import { MatSnackBar } from "@angular/material/snack-bar";
-import { levelConfig } from "src/app/model/levelConfiguration/level.configuration.model"
+import {Component, HostListener, OnInit, ViewChild} from '@angular/core';
+import {CheckSmellService} from "../../../services/check-smell/check-smell.service";
+import {ActivatedRoute} from "@angular/router";
+import {SuccessAlertComponent} from "../../../components/alert/success-alert/success-alert.component";
+import {AchievementAlertComponent} from "../../../components/alert/achivement-alert/achievement-alert.component";
+import {FailAlertComponent} from "../../../components/alert/fail-alert/fail-alert.component";
+import {Question} from "../../../model/question/question.model";
+import {ExerciseService} from "../../../services/exercise/exercise.service";
+import {UserService} from "../../../services/user/user.service";
+import {LeaderboardService} from "../../../services/leaderboard/leaderboard.service";
 
 @Component({
   selector: 'app-check-game-core',
@@ -17,16 +15,53 @@ import { levelConfig } from "src/app/model/levelConfiguration/level.configuratio
   styleUrls: ['./check-game-core-route.component.css']
 })
 export class CheckGameCoreRouteComponent implements OnInit {
-  ngOnInit(): void {
+  exerciseName!: string;
+
+  @ViewChild('achievementAlert') achievementAlert!: AchievementAlertComponent;
+  @ViewChild('failAlert') failAlert!: FailAlertComponent;
+  @ViewChild('successAlert') successAlert!: SuccessAlertComponent;
+
+  protected checkSmellService!: CheckSmellService;
+
+  constructor(
+    private exerciseService: ExerciseService,
+    private userService: UserService,
+    private route: ActivatedRoute,
+    private leaderboardService: LeaderboardService
+  ) {
+    this.exerciseName = this.route.snapshot.params['exercise'];
+    this.checkSmellService = new CheckSmellService(
+      this.exerciseService,
+      this.userService,
+      this.leaderboardService
+    )
   }
 
+  ngOnInit(): void {
+    this.checkSmellService.initQuestions(this.exerciseName);
+  }
+
+  async submitExercise(): Promise<void> {
+    this.checkSmellService.calculateScore();
+    this.checkSmellService.showResults(this.exerciseName).then(
+      () => {
+        if (this.checkSmellService.isExercisePassed()) {
+          this.successAlert.show();
+
+          if (this.userService.hasUserLevelledUp())
+            this.achievementAlert.show("Level Up!", "Congratulation, you have levelled up!");
+
+          if (this.userService.hasUserUnlockedBadge())
+            this.achievementAlert.show("Badge Unblocked!", "You have unlocked a new badge, view it on your profile page!");
+        } else {
+          this.failAlert.show();
+        }
+      }
+    );
 
 
 
+  }
 
-
-
-
-
-
+  protected readonly Math = Math;
 }
