+++ Test-Smells-Game/Projects/Frontend/src/app/routes/check-smell-game/check-smell-exercise-list/check-game-ex-list-route.component.html
@@ -10,17 +10,19 @@
   {{ 'Exercise server has problems' | translate }}
 </div>
 
-<ul class="list-group">
-  <li class="list-group-item" *ngFor="let item of exerciseConfigurations; let i = index">
-    <a [routerLink]="isExerciseEnabled(item.check_game_configuration.level) ? '/check-game/' + item.exerciseId.slice(0, -2) : null"
-       [style.textDecoration]="isExerciseEnabled(item.check_game_configuration.level) ? 'none' : 'line-through'"
-       [style.pointerEvents]="isExerciseEnabled(item.check_game_configuration.level) ? 'auto' : 'none'"
-       [style.color]="isExerciseEnabled(item.check_game_configuration.level) ? 'blue' : 'red'">
-      {{ item.exerciseId.slice(0, -2) }}
-      <span *ngIf="item.check_game_configuration.level === 1">⭐</span>
-      <span *ngIf="item.check_game_configuration.level === 2">⭐⭐</span>
-      <span *ngIf="item.check_game_configuration.level === 3">⭐⭐⭐</span>
-    </a>
-  </li>
-</ul>
+<div *ngIf="config">
+  <ul class="list-group">
+    <li class="list-group-item" *ngFor="let item of exerciseConfigurations; let i = index">
+      <a [routerLink]="isExerciseEnabled(item.check_game_configuration.level) ? '/check-game/' + item.exerciseId.slice(0, -2) : null"
+         [style.textDecoration]="isExerciseEnabled(item.check_game_configuration.level) ? 'none' : 'line-through'"
+         [style.pointerEvents]="isExerciseEnabled(item.check_game_configuration.level) ? 'auto' : 'none'"
+         [style.color]="isExerciseEnabled(item.check_game_configuration.level) ? 'blue' : 'red'">
+        {{ item.exerciseId.slice(0, -2) }}
+        <span *ngIf="item.check_game_configuration.level === 1">⭐</span>
+        <span *ngIf="item.check_game_configuration.level === 2">⭐⭐</span>
+        <span *ngIf="item.check_game_configuration.level === 3">⭐⭐⭐</span>
+      </a>
+    </li>
+  </ul>
+</div>
