+++ Test-Smells-Game/Projects/Frontend/src/app/routes/refactoring-game/refactoring-game-core/refactoring-game-core-route.component.html
@@ -1,23 +1,59 @@
-<app-refactoring-core [exerciseName]="exerciseName"
-                      #code [userCode]="userCode"
-                      [progressBarMode]="progressBarMode"
-                      #testing [testingCode]="testingCode"
-                      #output [shellCode]="shellCode"
-                      [exerciseSuccess]="exerciseSuccess"
-                      [refactoringWarning]="refactoringWarning"
-                      [smellNumber]="smellNumber"
-                      [refactoringResult]="refactoringResult"
-                      [originalCoverage]="originalCoverage"
-                      [refactoredCoverage]="refactoredCoverage"
-                      [smellNumberWarning]="smellNumberWarning"
-                      [exerciseConfiguration]="exerciseConfiguration"
-                      [smellList]="smellList"
-                      [methodList]="methodList"
-                      [smellDescriptions]="smellDescriptions"
-                      [solutionRepoRoute]="solutionRepoRoute"
-                      [compileCode]="compile"
-                      [submitExercise]="publishSolutionToLeaderboard"
-                      [submitIsEnabled]="submitIsEnabled"
-                      [finishButtonText]="'Save in solution repository'"
-                      [endButtonText]="'Solution Repository'"
-                      />
+<div class="container">
+  <h1>{{exerciseName}}</h1>
+  <div class="row">
+    <div class="col">
+    </div>
+    <div class="row">
+      <div class="col">
+        <app-refactoring-editor-card #code [editorHeader]="exerciseName+'.java'" [injectedCode]="refactoringService.userCode" [editable]="false"
+                                     [progressBarMode]="refactoringService.progressBarMode">
+        </app-refactoring-editor-card>
+      </div>
+
+      <mat-divider></mat-divider>
+
+      <div class="col" style="padding-top: 20px">
+        <app-refactoring-editor-card #testing [editorHeader]="exerciseName+'Test.java'" [injectedCode]="refactoringService.testingCode"
+                                     [progressBarMode]="refactoringService.progressBarMode">
+        </app-refactoring-editor-card>
+      </div>
+
+      <mat-divider></mat-divider>
+
+      <div style="margin-top: 20px">
+        <button type="button" class="btn btn-primary" (click)="compile()" [disabled]="refactoringService.progressBarMode == 'query'">{{ 'Compile' | translate }}</button>
+        <button style="margin-left: 20px;" type="button" class="btn btn-primary" [disabled]="submitIsEnabled()" (click)="refactoringService.publishSolution(exerciseName)">{{ 'Save in solution repository' | translate }}</button>
+      </div>
+
+      <mat-divider></mat-divider>
+
+      <div class="col" style="margin-top: 20px">
+        <app-refactoring-editor-card #output [editorHeader]="'Shell'" [injectedCode]="refactoringService.shellCode" [editable]="false"
+                                     [progressBarMode]="refactoringService.progressBarMode">
+        </app-refactoring-editor-card>
+      </div>
+
+      <app-refactoring-compilation-results></app-refactoring-compilation-results>
+
+      <div *ngIf="refactoringService.exerciseIsCompiledSuccessfully">
+        <app-refactoring-score
+          [score]="Math.abs(refactoringService.smellNumber - refactoringService.exerciseConfiguration.refactoring_game_configuration.smells_allowed)"
+          [smellNumber]="refactoringService.smellNumber"
+          [refactoringResult]="refactoringService.refactoringResult"
+          [originalCoverage]="refactoringService.originalCoverage"
+          [refactoredCoverage]="refactoringService.refactoredCoverage"
+          [smellNumberWarning]="refactoringService.smellNumberWarning"
+          [exerciseConfiguration]="refactoringService.exerciseConfiguration"
+        ></app-refactoring-score>
+
+        <app-refactoring-smell-list
+          [smellList]="refactoringService.smellList"
+          [smellDescriptions]="refactoringService.smellDescriptions"
+          [methodList]="refactoringService.methodList"
+        ></app-refactoring-smell-list>
+      </div>
+
+      <button type="button" class="btn btn-primary" routerLink="/refactor-game/leaderboard/{{exerciseName}}">{{ 'Solution Repository' | translate }}</button>
+    </div>
+  </div>
+</div>
